- Iterated on the public API in GO - Used LDB - Added a basic framework for importing external C/C++ functions (like getHeaderHashByBlockNumber) - Tested CGO compilation (interfacing go from C/C++)

+++ taraxa_vm/conflict_tracking/types.go
@@ -1,6 +1,6 @@
 package conflict_tracking
 
-type TxId = int
+type TxId = uint64
 type DUMMY = int
 
 const dummy DUMMY = 0
